% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_nardl.R
\name{ArchTest}
\alias{ArchTest}
\title{ARCH test}
\usage{
ArchTest(x, lags = 12, demean = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{lags}{positive integer number of lags}

\item{demean}{logical: If TRUE, remove the mean before computing the test statistic.}
}
\description{
Computes the Lagrange multiplier test for conditional heteroscedasticity of Engle (1982), as described by Tsay (2005, pp. 101-102).
}
\examples{

reg<-nardl(food~inf,fod,ic="aic",maxlag = 4,graph = TRUE,case=3)
x<-reg$selresidu
nlag<-reg$nl
ArchTest(x,lags=nlag)

}
