% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_nardl.R
\name{bp2}
\alias{bp2}
\title{LM test for serial correlation}
\usage{
bp2(object, nlags, fill = NULL, type = c("F", "Chi2"))
}
\arguments{
\item{object}{fitted lm model}

\item{nlags}{positive integer number of lags}

\item{fill}{starting values for the lagged residuals in the auxiliary regression. By default 0.}

\item{type}{Fisher or Chisquare statistics}
}
\description{
LM test for serial correlation
}
\examples{

reg<-nardl(food~inf,fod,ic="aic",maxlag = 4,graph = TRUE,case=3)
lm2<-bp2(reg$fits,reg$nl,fill=0,type="F")

}
