% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsq.R
\name{cumsq}
\alias{cumsq}
\alias{cumsq.nardl}
\alias{cumsq.default}
\title{Function cumsq}
\usage{
cumsq(x, ...)

\method{cumsq}{nardl}(x, ...)

\method{cumsq}{default}(x, k, n, ...)
}
\arguments{
\item{x}{is a object of the class \code{\link{nardl}} or a vector
with the recursive errors}

\item{...}{arguments passed to other methods. Currently ignored.}

\item{k}{is the estimated coefficients length}

\item{n}{is the recursive errors length}
}
\description{
Function cumsq
}
\examples{

reg<-nardl(food~inf,fod,ic="aic",maxlags = TRUE,graph = TRUE,case=3)
cumsq(reg)

# Or manually
e<-reg$rece
k<-reg$k
n<-reg$n
cumsq(x=e,k=k,n=n)

}
