% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fod.R
\docType{data}
\name{fod}
\alias{fod}
\title{Indian yearly data of inflation rate and percentage food import to total import}
\format{A data frame with 54 rows and 2 variables}
\usage{
data(fod)
}
\description{
The data frame \code{fod} contains the following variables:
\itemize{
\item{food: }{percentage food import to total import}
\item{inf: }{inflation rate}
\item{year: }{the year}
}
}
\keyword{datasets}
