% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmtest.R
\name{LMTest}
\alias{LMTest}
\alias{LMTest.nardl}
\alias{LMTest.summary.nardl}
\alias{bp2}
\title{LM test for serial correlation}
\usage{
LMTest(x, ...)

\method{LMTest}{nardl}(x, fill = NULL, type = c("F", "Chi2"), ...)

\method{LMTest}{summary.nardl}(x, ...)

bp2(object, nlags, fill = NULL, type = c("F", "Chi2"))
}
\arguments{
\item{x}{a \code{\link{nardl}} or \code{\link{summary.nardl}} object.}

\item{...}{arguments passed from and to other methods}

\item{fill}{starting values for the lagged residuals in the
auxiliary regression. By default 0.}

\item{type}{Fisher or Chisquare statistics}

\item{object}{fitted lm model}

\item{nlags}{positive integer number of lags}
}
\value{
a matrix with 1 row and 3 named columns, containing the
value for LM, the p value and the number of lags used.
}
\description{
This function performs a LM test for serial correlation. The function
\code{bp2()} is the workhorse, but the function \code{LMTest()} is
preferred as that function does the proper dispatching of methods.
}
\examples{

reg<-nardl(food~inf,fod,ic="aic",maxlags = TRUE,graph = TRUE,case=3)
LMTest(reg)
lm2<-bp2(reg$fit,reg$np,fill=0,type="F")

}
