% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nardl.R
\name{nardl.formula}
\alias{nardl.formula}
\title{formula}
\usage{
\method{nardl}{formula}(formula, p, q, data = list(), ...)
}
\arguments{
\item{formula}{food~inf}

\item{p}{scalar, autoregressive order for dependent variable}

\item{q}{scalar, autoregressive order for dependent variable}

\item{data}{the dataframe}

\item{...}{not used}
}
\description{
formula
}

