% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynmplier.R
\name{plotmplier}
\alias{plotmplier}
\title{Dynamic multiplier plot}
\usage{
plotmplier(model, np, k, h)
}
\arguments{
\item{model}{the fitted model}

\item{np}{the selected number of lags}

\item{k}{number of decomposed independent variables}

\item{h}{is the horizon over which multipliers will be computed}
}
\description{
Dynamic multiplier plot
}
\examples{

############################
# Dynamic multipliers plot
############################
# Load data
data(fod)
reg<-nardl(food~inf,p=4,q=4,fod,ic="aic",maxlags = FALSE,graph = TRUE,case=3)
plotmplier(reg,reg$np,1,10)

}
