% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nardl.R
\name{summary.nardl}
\alias{summary.nardl}
\title{summary}
\usage{
\method{summary}{nardl}(object, ...)
}
\arguments{
\item{object}{is the object of the function}

\item{...}{not used}
}
\description{
summary
}
