% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archtest.R
\name{ArchTest}
\alias{ArchTest}
\alias{ArchTest.nardl}
\alias{ArchTest.summary.nardl}
\alias{ArchTest.default}
\title{ARCH test}
\usage{
ArchTest(x, ...)

\method{ArchTest}{nardl}(x, demean = FALSE, ...)

\method{ArchTest}{summary.nardl}(x, ...)

\method{ArchTest}{default}(x, lags = 12, demean = FALSE, ...)
}
\arguments{
\item{x}{numeric vector, \code{\link{nardl}} or
\code{\link{summary.nardl}} object.}

\item{...}{arguments passed from or to other methods}

\item{demean}{logical: If TRUE, remove the mean before computing
the test statistic. This will be ignored when}

\item{lags}{positive integer number of lags. This will be ignored
when a \code{nardl} or \code{summary.nardl} is passed as \code{x}.}
}
\value{
a \code{htest} object.
}
\description{
Computes the Lagrange multiplier test for conditional
heteroscedasticity of Engle (1982), as described by
Tsay (2005, pp. 101-102).
}
\note{
This function is modeled after the \code{htest} class functions
like \code{\link[stats]{t.test}} and others. This also means that
when using the method on a \code{nardl} or \code{summary.nardl}
object, the name of the data displayed will always be ''residu''
}
\examples{

reg<-nardl(food~inf,fod,ic="aic",maxlags = TRUE,graph = TRUE,case=3)

ArchTest(reg)
ArchTest(reg, demean = TRUE)

reg.summ <- summary(reg)
ArchTest(reg)

# Manual calculation
x<-reg$selresidu
nlag<-reg$np
ArchTest(x,lags=nlag)

}
