% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nardl.R
\name{nardl}
\alias{nardl}
\title{method}
\usage{
nardl(x, ...)
}
\arguments{
\item{x}{a numeric design matrix for the model.}

\item{...}{not used}
}
\description{
method
}
\author{
Zaghdoudi Taha
}

