% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nardl.R
\name{nardl}
\alias{nardl}
\title{Nonlinear ARDL function}
\usage{
nardl(formula, data, p, q, ic = c("aic", "bic", "R2"), graph)
}
\arguments{
\item{formula}{food~inf or food~inf|I(inf^2)}

\item{data}{the dataframe}

\item{p}{max lags of dependent variable}

\item{q}{max lags of independent variables}

\item{ic}{: c("aic","bic","R2") criteria model selection}

\item{graph}{1 to show stability tests plot}
}
\description{
Nonlinear ARDL function
}

