% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nardl.R
\name{summary.nardl}
\alias{summary.nardl}
\title{Summary of a nardl model}
\usage{
\method{summary}{nardl}(object, ...)
}
\arguments{
\item{object}{is the object of the function}

\item{...}{not used}
}
\value{
an object of the S3 class \code{summary.nardl} with the
following components:
}
\description{
\code{summary} method for a \code{\link{nardl}} model.
}
