% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nardl.R
\name{summary.nardl}
\alias{summary.nardl}
\title{Summary of a nardl model}
\usage{
\method{summary}{nardl}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{nardl}}}

\item{...}{not used}
}
\value{
an object of the S3 class \code{summary.nardl} with the
following components:

 \itemize{
   \item icres: a matrix with the AIC, BIC, LL and R2 values
   \item sel: a \code{\link[stats]{summary.lm}} object with the actual fit
   \item np: the number of lags
   \item jbtest: a list with the results of the JB test
   \item lmtest: a list with the statistic nd p value related to
   the LM test.
   \item archtest: a list with the results of the arch test
   \item lres: a matrix with the long run coefficients and tests
   \item asymtest: a list with the statistic and p value of the
   long run asymmetry test.

 }
}
\description{
\code{summary} method for a \code{\link{nardl}} model.
}
\examples{
data(fod)
reg<-nardl(food~inf,fod,ic="aic",maxlags = TRUE,graph = FALSE,case=3)

summ.reg <- summary(reg)
summ.reg

}
